Separated songlist into user and admin area
Also added loading animation to songlist
